requirements.txt
tensorflow>=2.1
Pillow
numpy
opencv-python
